feat: i18n in index page
