+++ src/app/layout.tsx
@@ -1,10 +1,9 @@
 import { dir } from "i18next"
 import { Metadata } from "next"
-import { headers } from "next/headers"
 
 import "~/css/main.css"
+import { useAcceptLang } from "~/hooks/useAcceptLang"
 import { APP_DESCRIPTION, APP_NAME, SITE_URL } from "~/lib/env"
-import { fallbackLng } from "~/lib/i18n/settings"
 
 import Providers from "./providers"
 
@@ -28,17 +27,7 @@ export default function RootLayout({
 }: {
   children: React.ReactNode
 }) {
-  let lng = fallbackLng
-
-  let acceptLang = headers().get("accept-language")?.split(",")[0]
-  if (acceptLang === "zh-HK" || acceptLang === "zh-TW") {
-    acceptLang = "zh-TW"
-  } else {
-    acceptLang = acceptLang?.split("-")[0]
-  }
-  if (acceptLang) {
-    lng = acceptLang
-  }
+  const lng = useAcceptLang()
 
   return (
     <html lang={lng} dir={dir(lng)}>
