+++ src/components/common/FollowAllButton.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { useTranslation } from "next-i18next"
 import { useState } from "react"
